Update tmux conf and yank
Bracketed paste for tmux.

Fix middle-clicking in Vim.

Increase history limit so there is no need to ever redirect to a plain
file. Just hit M-/ and the output is there.

Add M-? to break out a pane into its own window. Replaces the "hole"
with a new pane to avoid mesing up layout.

M-x switches back to last window if needed. Not perfect: what I really
want is to switch to prior pane, _even_ if it is on a different window.

Make 'yank' work with various clipboard utils.